feat(cli, magefile): improve cli input handling, add local install mage target
Implemented boolean confirmation step for command line input gathering, introduced error handling for user confirmation, and a PR will only be submitted if the user confirms this action. Also, expanded the Magefile to include a new local installation target that enables building and installing the tool into the GOPATH/bin directory. This provides a path to test the application locally.